v1.4.2: Among Us Judge (2026.8.18) support
- Add 2026.8.18 to supported versions; bump to 1.4.2
- Fix Judge API renames (PlayerVoteArea.PlayerId, MeetingHud.MeetingStates, RpcClearVote(PlayerId))
- Judge role: durable color fallback, host assignment pool, GameSummary shows real roles for the dead
- Anti-cheat: catch fake-Judge overrule (QueueOverruleVotes) + non-host VotingComplete; dedup unknown-RPC popup
- Log Judge overrule (encrypted) to the event log
- Fix NPEs: null-safe GetBehaviourPrefab/IsImpostor, PlayerInfoDisplay Data.Role/targetData guards
- Guardian Angel now sees roles like other dead players

+++ src/Commands/ForceSkipCommand.cs
@@ -31,7 +31,7 @@ internal sealed class ForceSkipCommand : BaseCommand
         {
             foreach (var client in AmongUsClient.Instance.allClients)
             {
-                MeetingHud.Instance.RpcClearVote(client.Id);
+                MeetingHud.Instance.RpcClearVote((byte)client.Id);
             }
             MeetingHud.Instance.RpcClose();
         }
